LayoutTests/http/tests/referrer-policy-script/origin-when-cross-origin/cross-origin-http-http.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests the behavior of origin-when-cross-origin referrer policy when cross origin.");
jsTestIsAsync = true;

function checkReferrer(value) {
    referrer = value;
    // Should be the origin, not the full URL, because we are cross-origin.
    shouldBeEqualToString("referrer", "http://127.0.0.1:8000/");
    finishJSTest();
}
</script>
<script src="http://localhost:8000/referrer-policy/resources/script.php" referrerpolicy="origin-when-cross-origin"></script>
</body>
</html>
